Big update to logic flow
Breakout board design for DACs significantly updated to have in-series
impedance and better ground shepherd plane under logic lines. Focused on
design for AD7945 for 2 MHz updates.

FPGA design code expands memory to allow ~69 seconds of waveform data,
external logic triggers to run each channel, and an op-code for a "free
run" mode for the DAC in which it continues to repeat the first waveform
in memory forever.