Major update! Full c++ and FPGA logic control
This is a huge update. The FPGA and C++ program both now have features
to do logic sequencing. This includes logic switching on all LOGIC
outputs, except when the ADCs throw a warning that a voltage is low. It
also includes synchronous triggering of waveforms during logic
sequencing. There are two input triggers, one for DACs alone and one for
logic sequencing which may include DACs. The C++ program has a very
simple parser for logic steps as well as a method to pre-upload waveform
data by the channel.