Update to logic sequencing
Both VHDL and C++ control code both now allow logic sequencing with
external or internal triggers for DACs and Logic sequences. This
includes being able to choose when the ADC interlock system is active
for TTLs 0 and 1 based on ADCs 0 and 1 respectively. The command console
is rewritten to allow calling functions such as for loading waveform
data, allowing an experiment to be defined without the command console
and simply writing a function that calls potentially many waveforms and
logic sequences with a final run command.